Refactor role and user management in controllers and services
- Removed unused Request import from MenuController.
- Introduced RoleService and UserService to encapsulate role and user modification logic.
- Updated RoleController and UserController to utilize the new services for checking modifiability of roles and users.
- Enhanced role deletion logic to assign default roles to users without roles.
- Added new methods in RoleService and UserService for role and user modifiability checks.
- Updated database seeders to include a default role and user for consistency.

+++ app/Http/Controllers/RoleController.php
@@ -2,8 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-
 use Inertia\Inertia;
 
 use Spatie\Permission\Models\Role;
@@ -11,8 +9,17 @@ use Spatie\Permission\Models\Role;
 use App\Http\Requests\Role\StoreRoleRequest;
 use App\Http\Requests\Role\UpdateRoleRequest;
 
+use App\Services\RoleService;
+
 class RoleController extends Controller
 {
+    protected $roleService;
+
+    public function __construct(RoleService $roleService)
+    {
+        $this->roleService = $roleService;
+    }
+
     public function index()
     {
         $roles = Role::all();
@@ -29,13 +36,27 @@ class RoleController extends Controller
 
     public function update(UpdateRoleRequest $request, Role $role)
     {
+
+        if (!$this->roleService->veryIfRoleIsModifiable($role->name)) {
+            return to_route('roles.index')->with('error', 'Role cannot be modified');
+        }
+
         $role->update($request->all());
         return to_route('roles.index')->with('success', 'Role updated successfully');
     }
 
     public function destroy(Role $role)
     {
+        if (!$this->roleService->veryIfRoleIsModifiable($role->name)) {
+            return to_route('roles.index')->with('error', 'Role cannot be deleted');
+        }
+
+        $usersWithRole = $role->users;
+        
         $role->delete();
+
+        $this->roleService->setDefaultRole($usersWithRole);
+
         return to_route('roles.index')->with('success', 'Role deleted successfully');
     }
 }
